pyproject.toml
[project]
name = "alfred-local-workbench"
version = "1.0.0"
description = "Local-first Python orchestrator and FastAPI bridge for Alfred CLI"
authors = [{ name = "Johannes", email = "johannes@aibyjohannes.com" }]
requires-python = ">=3.12,<4"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "fastapi>=0.115.0,<0.116",
    "uvicorn[standard]>=0.32.0,<0.33",
    "pydantic>=2.9.0,<3",
    "ddgs>=9.5.4,<10",
    "smolagents[openai]>=1.0.0,<2",
    "python-multipart>=0.0.6,<1",
]

[project.optional-dependencies]
transcribe = [
    "faster-whisper>=1.0.0,<2",
]
nemo = [
    "nemo-toolkit[asr]>=2.0.0,<3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.24.0,<0.25",
    "pytest-cov>=6.0.0,<7",
    "black>=24.10.0,<25",
    "ruff>=0.7.0,<0.8",
    "mypy>=1.13.0,<2",
]

[tool.hatch.build.targets.wheel]
packages = ["llm", "prompts", "scripts", "transcription"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
